pyproject.toml
[project]
name = "ddd-ai-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.56.0",
    "pydantic==2.12.5",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]
